+++ cypress/cypress/component/Typeahead/AsyncTypeaheadInput.cy.tsx
@@ -282,3 +282,36 @@ it("empty label", () => {
   cy.get(`#${name}`).clear().click().type(name);
   cy.get(".dropdown-menu > .dropdown-item").should("have.text", emptyLabel);
 });
+
+it("placeholder", () => {
+  const { simpleOptions } = generateOptions();
+  const name = faker.random.alpha(10);
+  const placeholder = faker.random.words(5);
+
+  cy.mount(
+    <Form
+      onSubmit={cy.spy().as("onSubmitSpy")}
+      defaultValues={{
+        [name]: simpleOptions,
+      }}
+    >
+      <AsyncTypeaheadInput
+        multiple
+        name={name}
+        label={name}
+        queryFn={async (query) =>
+          await fetchMock(
+            simpleOptions.map((x) => {
+              return { label: x, value: x };
+            }),
+            query,
+            false,
+          )
+        }
+        placeholder={placeholder}
+      />
+      <input type="submit" />
+    </Form>,
+  );
+  cy.get(`#${name}`).should("have.attr", "placeholder", placeholder);
+});