+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `emptyLabel` property to `TypeaheadInput` components
+- Added `placeholder` property to `TypeaheadInput` components
 
 ## [2.2.2] - 2023-12-01