+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2023-12-18
+
 ### Added
 
 - Added `step` property to `Input` component
@@ -354,7 +356,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.3.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.4.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -379,6 +381,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[2.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.2.2...2.3.0
 [2.2.2]: https://github.com/neolution-ch/react-hook-form-components/compare/2.2.1...2.2.2
 [2.2.1]: https://github.com/neolution-ch/react-hook-form-components/compare/2.2.0...2.2.1